Fixes Uploader misses some blocks
If the uploader fails to upload a block after trying for 800ms it means
that the block is not owned by the bot profile or that the location is
invalid. Sometimes the 80ms were not enough to ensure this and therefore
blocks were said to be invalid but the problem was only lag.